Add: initial setup for integration tests of backend servers
Refactored program.cs in both backend projects to be available for import through web application builder in integrationtests, added inmemory db to both test server instances
